Extract functions into their own modules
To make the main.py file less bloated, some methods got extracted into their own python modules.
